hameg-hmo: Change the way vdiv and timebase are saved.
The current vertical division setting (per channel) and the timebase are stored
as a floating point number. This is suboptimal since clients expect us to send
this information to them in form of a rational number.

Store only the index of the current setting since all the supported settings are
already stored inside of an array.

+++ hardware/hameg-hmo/protocol.h
@@ -69,7 +69,7 @@ struct scope_config {
 struct analog_channel_state {
 	int coupling;
 
-	float vdiv;
+	int vdiv;
 	float vertical_offset;
 
 	gboolean state;
@@ -80,7 +80,7 @@ struct scope_state {
 	gboolean *digital_channels;
 	gboolean *digital_pods;
 
-	float timebase;
+	int timebase;
 	float horiz_triggerpos;
 
 	int trigger_source;
